Added file fuzzy match and refactored
